move term stuff to its own module
